Propagate to FRANKLIN:

Change 85184 by dkarp@dkarp-imap on 2008/02/23 12:11:42

        Fix a ?: compiler complaint.

http://bugzilla.zimbra.com/show_bug.cgi?id=22577

Copied from Perforce
 Change: 85605

+++ src/java/com/zimbra/cs/mailbox/calendar/cache/FullInstanceData.java
@@ -118,7 +118,7 @@ public class FullInstanceData extends InstanceData {
         mRecurrenceId = recurrenceId; mSequence = sequence; mDtStamp = dtStamp;
         mOrganizer = organizer; mIsOrganizer = isOrganizer;
         mAttendees = attendees;
-        mNumAttendees = attendees != null ? attendees.size() : null;
+        mNumAttendees = attendees != null ? (Integer) attendees.size() : null;
         mSummary = summary; mLocation = location; mDescription = description; mFragment = fragment;
         mIsAllDay = isAllDay;
         mStatus = status; mPriority = priority; mClassProp = classProp;
